ineed a library... structuring pure functions in index
